Committing firt working version of the client without creds
This commit sets up the main file in the cmd directory as well as
adding the tools functions with the stubs for many of the intended
functions and the one working function right now to pull stock
prices. As of this commit, you need to hard-code your API creds to
access the polygon.io stock API. The next commit aims to address
hard-coding API creds to remove that depenency.