stockClient.go,TechAnalysis.go: fix bug causing crypto prices to show up a day behind and fix calls to crypto in polygon to not return blank values
This commit addresses bugs found here:
https://github.com/khrystoph/portfoliotools/issues/3
https://github.com/khrystoph/portfoliotools/issues/2

polygon now gets proper calculations because the call using the ticker has been removed
and it just calculates everything for every ticker. Right now, the program only accepts
a single ticker, but there are plans to support multiple tickers in the future as this
tool will eventually serve as a backfill tool. Alpaca also now returns the most recent
day of crypto from logic added to exclude the day subtraction from the data pull. The
result is the previous day's data is pulled instead of two days ago for crypto, while
maintaining the previous day for equities. Previously, equities would throw an error
on alpaca if using the current date as the end date, particularly if it was a weekend
day. This should now be completely resolved at each end.