bug: calculateRiskRange low value sometimes goes negative and fixes the previous failed fix attempt
This fix addresses a previous attempt to prevent the vol-adjusted ranges from going negative (because negative price in equities is not possible). The previous fix tried to use math.Max, but this did not result in an outcome where these values are properly set. Changing the logic to simple "is rrlow > zero?" corrects this